feat(amp): starter and solution
This commit adds the starter and solution code for a one-day (2-3 hour)
AMP beginner course. During the course students will complete 2 - 3
practicals, building out a single AMP page (or a simple app containing
AMP pages). This code is taken from the AMP foundations Google codelab
and will be adapted to the requirements of this course.

Change-Id: I75fc4e447a648eb50c6b13ad6cc02c4069ed3048