refactor(all): added intermediate solution
This commit adds a solution for the first half of the lab (step 7). It
also removes the starting HTML and CSS files from the starter folder so
that students can practice writing this for themselves. TODO - finish
writing instructions for the second half of the lab and update the
solution code as needed.

Change-Id: I870c15efe81181124addc325e99a242e5571fc65