update(all): update basic solution + amp-accordian
This commit combines the solutions from labs 1 and 2 from the first course (basic amp) into one 'basic-solution'. It also updates the code to use the css from the Simple Article AMPStart template. TODO - update lab instructions to match new look.

This commit adds an intermediate solution folder. The code is currently the same as the basic-solution code, but with an amp-accordian component added to the sidebar. TODO - build out the intermediate solution with the rest of the components in the training outline