feat(products.html): add products page
This commit adds a products page inside the pages folder and some images of the new various products offered by Chico's Cheese  Bicycles. The products page uses amp-list to fetch a list of products from a JSON endpoint. It uses amp-bind to let the user filter the products by type and sort the products by price. TODO: fix a bug that removes the filter and shows all products when the user changes the sort-by selector.