feat(all): convert to glitch and sorting/filtering
This commit refactors the sample app to work on Glitch so that we can take advantage of their HTTPS hosting. The project on Glitch is currently hosted here: https://glitch.com/edit/#!/seemly-metal. Filtering and sorting is working on the products page. TODO - add link to products page in the sidebar. TODO - add product details page